requirement.txt
absl-py==2.1.0
aniso8601==9.0.1
astunparse==1.6.3
bcrypt==4.2.1
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.0
certifi==2024.8.30
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.9
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flatbuffers==25.1.24
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.4
keras==3.8.0
keyring==24.3.1
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.5.0
msgpack==1.1.0
mysqlclient==2.2.5
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pexpect==4.9.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
protobuf==5.29.3
ptyprocess==0.7.0
Pygments==2.19.1
PyJWT==2.10.0
pyproject_hooks==1.2.0
pytz==2024.2
pywin32-ctypes==0.2.3
RapidFuzz==3.10.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==75.8.0
shellingham==1.5.4
six==1.16.0
SQLAlchemy==2.0.36
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tomlkit==0.13.2
trove-classifiers==2024.10.13
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.6
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
